requirements.txt
httpx
dnspython
Jinja2
boofuzz
requests
fastapi
uvicorn